chore(ansible-playbooks): limit unix sock path by changing location
Unix socket path length is limited to 108 characters on Linux. This
means that Firecracker could not be able to spawn VMs if the provided
path length is bigger than that. To allow the socket path to be of a
reasonable length, put it directly under /tmp, and use a combination
of the run id and the machine name to define the file name. This
should be ok for majority of applications.

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>