chore(ansible-playbooks): revert to v0.3.2 for ubuntu-6.5-aarch64
This is a temporary solution to make ubuntu-6.5 tests available on
aarch64. Unfortunately, the newer images don't work, due to the fact
that internal VM networking is not properly configured: specifically,
the VM is able to ask for networking configuration to the external
dnsmasq service, but the configuration is, for some reason, re-set
after OS booting. While we investigate this issue, use the working
old image. Furtunately, the old image has a recent enough cmake
version (3.27.4).

Signed-off-by: Leonardo Di Giovanna <leonardodigiovanna1@gmail.com>